fix: rename migration revision ID to avoid conflict and use exec_driver_sql
The revision ID a1b2c3d4e5f6 conflicted with the existing
rename_scaling_group migration. Changed to ccf8ae5c90fe.
Also switched sa.text() to conn.exec_driver_sql() to avoid
JSON colons being parsed as SQLAlchemy bind parameters.
